+++ pkg/apis/kubeone/config/config.go
@@ -17,13 +17,14 @@ limitations under the License.
 package config
 
 import (
+	"fmt"
 	"io"
 	"os"
 	"os/exec"
+	"reflect"
 
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
 
 	kubeoneapi "k8c.io/kubeone/pkg/apis/kubeone"
 	kubeonescheme "k8c.io/kubeone/pkg/apis/kubeone/scheme"
@@ -34,6 +35,7 @@ import (
 	terraformv1beta2 "k8c.io/kubeone/pkg/terraform/v1beta2"
 
 	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/yaml"
 )
 
 const (
@@ -168,7 +170,7 @@ func DefaultedV1Beta1KubeOneCluster(versionedCluster *kubeonev1beta1.KubeOneClus
 	}
 
 	// Check for deprecated fields/features for a cluster
-	checkClusterForDeprecations(*internalCluster, logger)
+	checkClusterFeatures(*internalCluster, logger)
 
 	return internalCluster, nil
 }
@@ -206,13 +208,13 @@ func DefaultedV1Beta2KubeOneCluster(versionedCluster *kubeonev1beta2.KubeOneClus
 	}
 
 	// Check for deprecated fields/features for a cluster
-	checkClusterForDeprecations(*internalCluster, logger)
+	checkClusterFeatures(*internalCluster, logger)
 
 	return internalCluster, nil
 }
 
 // SetKubeOneClusterDynamicDefaults sets the dynamic defaults for a given KubeOneCluster object
-func SetKubeOneClusterDynamicDefaults(cfg *kubeoneapi.KubeOneCluster, credentialsFile []byte) error {
+func SetKubeOneClusterDynamicDefaults(cluster *kubeoneapi.KubeOneCluster, credentialsFile []byte) error {
 	// Parse the credentials file
 	credentials := make(map[string]string)
 	err := yaml.Unmarshal(credentialsFile, &credentials)
@@ -222,7 +224,17 @@ func SetKubeOneClusterDynamicDefaults(cfg *kubeoneapi.KubeOneCluster, credential
 
 	// Source cloud-config from the credentials file if it's present
 	if cc, ok := credentials["cloudConfig"]; ok {
-		cfg.CloudProvider.CloudConfig = cc
+		cluster.CloudProvider.CloudConfig = cc
+	}
+	// Source csi-config from the credentials file if it's present
+	if cc, ok := credentials["csiConfig"]; ok {
+		cluster.CloudProvider.CSIConfig = cc
+	}
+
+	if ra, ok := credentials["registriesAuth"]; ok {
+		if err := setRegistriesAuth(cluster, ra); err != nil {
+			return fmt.Errorf("failed to parse registriesAuth from credentials file: %w", err)
+		}
 	}
 
 	// check if the container log max size is set, if not set it to the of 100Mi
@@ -231,7 +243,47 @@ func SetKubeOneClusterDynamicDefaults(cfg *kubeoneapi.KubeOneCluster, credential
 	}
 
 	// Default the AssetsConfiguration internal API
-	cfg.DefaultAssetConfiguration()
+	cluster.DefaultAssetConfiguration()
+
+	return nil
+}
+
+func setRegistriesAuth(cluster *kubeoneapi.KubeOneCluster, buf string) error {
+	var (
+		containerdConfig kubeonev1beta2.ContainerRuntimeContainerd
+		typeMeta         runtime.TypeMeta
+	)
+
+	if err := yaml.Unmarshal([]byte(buf), &typeMeta); err != nil {
+		return err
+	}
+
+	if typeMeta.APIVersion != kubeonev1beta2.SchemeGroupVersion.String() {
+		return fmt.Errorf("only %q apiVersion is supported in registriesAuth", kubeonev1beta2.SchemeGroupVersion.String())
+	}
+
+	containerdConfigKind := reflect.TypeOf(containerdConfig).Name()
+	if typeMeta.Kind != containerdConfigKind {
+		return fmt.Errorf("only %q kind is supported in registriesAuth", containerdConfigKind)
+	}
+
+	if err := yaml.Unmarshal([]byte(buf), &containerdConfig); err != nil {
+		return err
+	}
+
+	if cluster.ContainerRuntime.Containerd == nil {
+		return fmt.Errorf(".ContainerRuntime.Containerd should be set")
+	}
+
+	if cluster.ContainerRuntime.Containerd.Registries == nil {
+		cluster.ContainerRuntime.Containerd.Registries = map[string]kubeoneapi.ContainerdRegistry{}
+	}
+
+	for registryName, registryInfo := range containerdConfig.Registries {
+		internalRegistry := cluster.ContainerRuntime.Containerd.Registries[registryName]
+		internalRegistry.Auth = (*kubeoneapi.ContainerdRegistryAuthConfig)(registryInfo.Auth)
+		cluster.ContainerRuntime.Containerd.Registries[registryName] = internalRegistry
+	}
 
 	return nil
 }
@@ -241,9 +293,13 @@ func isDir(dirname string) bool {
 	return statErr == nil && stat.Mode().IsDir()
 }
 
-// checkClusterForDeprecations with check clusters for usage of deprecated fields, flags etc. and print a warning if any are found
-func checkClusterForDeprecations(c kubeoneapi.KubeOneCluster, logger logrus.FieldLogger) {
+// checkClusterFeatures checks clusters for usage of alpha and deprecated fields, flags etc. and print a warning if any are found
+func checkClusterFeatures(c kubeoneapi.KubeOneCluster, logger logrus.FieldLogger) {
 	if c.Features.PodSecurityPolicy != nil && c.Features.PodSecurityPolicy.Enable {
 		logger.Warnf("PodSecurityPolicy is deprecated and will be removed with Kubernetes 1.25 release")
+	}
+	if c.CloudProvider.Nutanix != nil {
+		logger.Warnf("Nutanix support is considered as alpha, so the implementation might be changed in the future")
+		logger.Warnf("Nutanix support is planned to graduate to beta/stable in KubeOne 1.5+")
 	}
 }